% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{gpt_params}
\alias{gpt_params}
\title{Estimate GPT Total Parameters}
\usage{
gpt_params(vocab_size, n_layer, n_head, n_embed, block_size)
}
\arguments{
\item{vocab_size}{number of vocabulary tokens}

\item{n_layer}{number of transformer layers}

\item{n_head}{number of transfomer heads per layer}

\item{n_embed}{size of feature (embedding) dimension}

\item{block_size}{max context size for position encodings}
}
\description{
Estimate GPT Total Parameters
}
\examples{
gpt_params(50257, 12, 12, 768, 1024) # gpt2

}
