% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{gpt_size}
\alias{gpt_size}
\title{Estimate GPT Total Parameters}
\usage{
gpt_size(vocab_size, n_layer, n_head, n_embed, block_size)
}
\arguments{
\item{vocab_size}{number of vocabulary tokens}

\item{n_layer}{number of transformer layers}

\item{n_head}{number of transfomer heads per layer}

\item{n_embed}{size of feature (embedding) dimension}

\item{block_size}{max context size for position encodings}
}
\description{
Estimate GPT Total Parameters
}
\examples{
gpt_size(50257, 12, 12, 768, 1024) # gpt2
gpt_size(50257, 24, 16, 1024, 1024) # gpt2-med
gpt_size(50257, 36, 20, 1280, 1024) # gpt2-large
gpt_size(50257, 48, 25, 1600, 1024) # gpt2-xl

}
