% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llama.r
\name{layer_llama_rmsnorm}
\alias{layer_llama_rmsnorm}
\title{LLaMA RMS Norm Layer}
\usage{
layer_llama_rmsnorm(
  object,
  eps = 1e-06,
  ...,
  block_id = NULL,
  feeds_into = NULL
)
}
\description{
LLaMA RMS Norm Layer
}
