% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llama.r
\name{llama_hidden_dim}
\alias{llama_hidden_dim}
\title{Calculate hidden dim for feedforward network}
\usage{
llama_hidden_dim(output_dim, multiple_of)
}
\description{
Rounds 8/3*output_dim to the next multiple of multiple_of
}
