% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llama.r
\name{llama_model}
\alias{llama_model}
\title{Create LLaMA model}
\usage{
llama_model(params, max_context_size = 2048L, use_cache = FALSE)
}
\arguments{
\item{params}{list of params}
}
\description{
Create LLaMA model
}
