% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotary-embedding.r
\name{rope}
\alias{rope}
\title{ROtational Position Embeddings}
\usage{
rope(x, rots)
}
\arguments{
\item{x}{inputs}

\item{rots}{precomputed rotational matrix from \code{rotation_matrix}}
}
\description{
ROtational Position Embeddings
}
