% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotary-embedding.r
\name{rope_matrix}
\alias{rope_matrix}
\title{Rotation Matrix for RoPE}
\usage{
rope_matrix(seqlen, feature_dim, theta = 10000)
}
\arguments{
\item{seqlen}{size of sequence}

\item{feature_dim}{size of features}
}
\description{
Rotation Matrix for RoPE
}
